Added new extensions UseExceptionHandler & UseDeveloperExceptionPage
