+++ samples/StatusCodePagesSample/Startup.cs
@@ -17,7 +17,7 @@ namespace StatusCodePagesSample
     {
         public void Configure(IApplicationBuilder app)
         {
-            app.UseErrorPage();
+            app.UseDeveloperExceptionPage();
             app.UseStatusCodePages(); // There is a default response but any of the following can be used to change the behavior.
 
             // app.UseStatusCodePages(context => context.HttpContext.Response.SendAsync("Handler, status code: " + context.HttpContext.Response.StatusCode, "text/plain"));
